Moves reference star calibration to OOD
