+++ snat_sim/reference_stars.py
@@ -16,183 +16,306 @@ import numpy as np
 import pandas as pd
 from astropy.table import Table
 
+from . import constants as const
 from ._data_paths import data_paths
+from .models import PWVModel
 
 Numeric = Union[int, float]
 
 
-def get_available_types() -> List[str]:
-    """Return the spectral types available on disk
+class ReferenceStar:
+    """Representation of spectral data from the Goettingen Spectral Library"""
 
-    Returns:
-        A list fo spectral types
-    """
+    def __init__(self, spectral_type: str) -> None:
+        """Load a spectrum for the given spectral type
 
-    return sorted(f.stem for f in data_paths.stellar_flux_dir.glob('*.txt'))
+        Flux values are returned in phot/cm2/s/angstrom and are index by
+        wavelength values in Angstroms.
+
+        Args:
+            spectral_type: Spectral type (e.g., G2)
+        """
+
+        self.spectral_type = spectral_type
+        if spectral_type not in self.get_available_types():
+            raise ValueError(f'Data for spectral type "{spectral_type}" is not available.')
+
+        # Load spectra for different spectral types
+        path = next(data_paths.stellar_spectra_dir.glob(spectral_type + '*.fits'))
+        self._spectrum = self._read_stellar_spectra_path(path)
+
+    @property
+    def flux(self) -> np.array:
+        """Flux values of the spectrum"""
+
+        return self._spectrum.to_numpy()
+
+    @property
+    def wavelength(self) -> np.array:
+        """Wavelength values of the spectrum"""
+
+        return self._spectrum.index.to_numpy()
+
+    def to_pandas(self) -> pd.Series:
+        """Return the spectral data as a ``pandas.Series`` object"""
+
+        return self._spectrum.copy()
+
+    @staticmethod
+    def get_available_types() -> List[str]:
+        """Return the spectral types available on disk
+
+        Returns:
+            A list fo spectral types
+        """
+
+        return sorted(f.stem for f in data_paths.stellar_flux_dir.glob('*.txt'))
+
+    @staticmethod
+    def _read_stellar_spectra_path(fpath: Union[str, Path]) -> pd.Series:
+        """Load fits file with stellar spectrum from phoenix
+
+        Fits files can be downloaded from:
+          http://phoenix.astro.physik.uni-goettingen.de/?page_id=15
+
+        converts from egs/s/cm2/cm to phot/cm2/s/nm using
+          https://hea-www.harvard.edu/~pgreen/figs/Conversions.pdf
+
+        Flux values are returned in phot/cm2/s/angstrom and are index by
+        wavelength values in Angstroms.
+
+        Args:
+            fpath: Path of the file to read
+
+        Returns:
+            Flux values as a pandas Series
+        """
+
+        # Load spectral data
+        with fits.open(fpath) as infile:
+            spec = infile[0].data
+
+        # Load data used to convert spectra to new units
+        # noinspection SpellCheckingInspection
+        with fits.open(fpath.parent / 'WAVE_PHOENIX-ACES-AGSS-COND-2011.fits') as infile:
+            lam = infile[0].data  # angstroms
+
+        angstroms_per_cm = 1e8
+        conversion_factor = 5.03 * 10 ** 7  # See https://hea-www.harvard.edu/~pgreen/figs/Conversions.pdf
+        ergs_per_photon = conversion_factor * lam
+
+        # Evaluate unit conversion
+        spec /= angstroms_per_cm  # ergs/s/cm2/cm into ergs/s/cm2/Angstrom
+        spec *= ergs_per_photon  # into phot/cm2/s/angstrom
+
+        indices = (lam >= 3000) & (lam <= 12000)
+        return pd.Series(spec[indices], index=lam[indices])
+
+    @lru_cache()  # Cache I/O
+    def get_ref_star_dataframe(self) -> pd.DataFrame:
+        """Retrieve PWV values to use as reference values
+
+        Args:
+            spectral_type: Type of reference star (Default 'G2')
+
+        Returns:
+            A DataFrame indexed by PWV with columns for flux
+        """
+
+        rpath = data_paths.stellar_flux_dir / f'{self.spectral_type}.txt'
+        if not rpath.exists():
+            raise ValueError(
+                f'Flux data not available for specified star {self.spectral_type}. '
+                f'Could not find: {rpath}')
+
+        band_names = [f'lsst_hardware_{b}' for b in 'ugrizy']
+        column_names = ['PWV'] + band_names
+        reference_star_flux = pd.read_csv(
+            rpath,
+            sep='\\s',
+            header=None,
+            names=column_names,
+            comment='#',
+            index_col=0,
+            engine='python'
+        )
+
+        for band in band_names:
+            band_flux = reference_star_flux[f'{band}']
+            reference_star_flux[f'{band}_norm'] = band_flux / band_flux.loc[0]
+
+        return reference_star_flux
+
+    def flux(self, band: str, pwv: Union[Numeric, np.array]) -> np.ndarray:
+        """Return the reference star flux values
+
+        Args:
+            band: Band to get flux for
+            pwv: PWV values to get magnitudes for
+
+        Returns:
+            The normalized flux at the given PWV value(s)
+        """
+
+        reference_star_flux = self.get_ref_star_dataframe()
+        if np.any(
+                (pwv < reference_star_flux.index.min()) |
+                (pwv > reference_star_flux.index.max())):
+            raise ValueError('PWV is out of range')
+
+        norm_flux = reference_star_flux[band]
+        return np.interp(pwv, norm_flux.index, norm_flux)
+
+    def norm_flux(self, band: str, pwv: Union[Numeric, np.array]) -> np.ndarray:
+        """Return the normalized reference star flux values
+
+        Args:
+            band: Band to get flux for
+            pwv: PWV values to get magnitudes for
+
+        Returns:
+            The normalized flux at the given PWV value(s)
+        """
+
+        reference_star_flux = self.get_ref_star_dataframe()
+        if np.any(
+                (pwv < reference_star_flux.index.min()) |
+                (pwv > reference_star_flux.index.max())):
+            raise ValueError('PWV is out of range')
+
+        norm_flux = reference_star_flux[f'{band}_norm']
+        return np.interp(pwv, norm_flux.index, norm_flux)
 
 
-def _read_stellar_spectra_path(fpath: Union[str, Path]) -> pd.Series:
-    """Load fits file with stellar spectrum from phoenix
+class ReferenceCatalog:
+    """A rudimentary implementation of a reference star catalog"""
 
-    Fits files can be downloaded from:
-      http://phoenix.astro.physik.uni-goettingen.de/?page_id=15
+    def __init__(self, *spectral_types: str) -> None:
+        """Create a reference star catalog composed of the given spectral types
 
-    converts from egs/s/cm2/cm to phot/cm2/s/nm using
-      https://hea-www.harvard.edu/~pgreen/figs/Conversions.pdf
+        Args:
+            *spectral_types: Spectral types for the catalog (e.g., 'G2', 'M5', 'K2')
+        """
 
-    Flux values are returned in phot/cm2/s/angstrom and are index by
-    wavelength values in Angstroms.
+        if not spectral_types:
+            raise ValueError('Must specify at least one spectral type for the catalog.')
 
-    Args:
-        fpath: Path of the file to read
+        self.spectral_types = spectral_types
+        self.spectra = (ReferenceStar(st) for st in spectral_types)
 
-    Returns:
-        Flux values as a pandas Series
-    """
+    def average_norm_flux(self, band: str, pwv: Union[Numeric, Collection, np.ndarray], ) -> np.ndarray:
+        """Return the average normalized reference star flux
 
-    # Load spectral data
-    with fits.open(fpath) as infile:
-        spec = infile[0].data
+        Args:
+            band: Band to get flux for
+            pwv: PWV values to get magnitudes for
 
-    # Load data used to convert spectra to new units
-    # noinspection SpellCheckingInspection
-    with fits.open(fpath.parent / 'WAVE_PHOENIX-ACES-AGSS-COND-2011.fits') as infile:
-        lam = infile[0].data  # angstroms
+        Returns:
+            The normalized flux at the given PWV value(s)
+        """
 
-    angstroms_per_cm = 1e8
-    conversion_factor = 5.03 * 10 ** 7  # See https://hea-www.harvard.edu/~pgreen/figs/Conversions.pdf
-    ergs_per_photon = conversion_factor * lam
+        return np.average([s.norm_flux(band, pwv) for s in self.spectra], axis=0)
 
-    # Evaluate unit conversion
-    spec /= angstroms_per_cm  # ergs/s/cm2/cm into ergs/s/cm2/Angstrom
-    spec *= ergs_per_photon  # into phot/cm2/s/angstrom
+    def divide_ref_from_lc(self, lc_table: Table, pwv: Union[Numeric, np.ndarray]) -> Table:
+        """Divide reference flux from a light-curve
 
-    indices = (lam >= 3000) & (lam <= 12000)
-    return pd.Series(spec[indices], index=lam[indices])
+        Recalibrate flux values using the average change in flux of a collection of
+        reference stars.
 
+        Args:
+            lc_table: Astropy table with columns ``flux`` and ``band``
+            pwv: PWV value to subtract reference star for
 
-def get_stellar_spectra(spectral_type: str) -> pd.Series:
-    """Load spectrum for given spectral type
+        Returns:
+            A modified copy of ``lc_table``
+        """
 
-    Flux values are returned in phot/cm2/s/angstrom and are index by
-    wavelength values in Angstroms.
+        if isinstance(pwv, Collection):
+            if len(pwv) != len(lc_table):
+                raise ValueError('PWV must be a float or have the same length as ``lc_table``')
 
-    Args:
-        spectral_type: Spectral type (e.g., G2)
+            pwv = np.array(pwv)
 
-    Returns:
-        Flux values as a pandas Series
-    """
+        else:
+            pwv = np.full(len(lc_table), pwv)
 
-    # Load spectra for different spectral types
-    path = next(data_paths.stellar_spectra_dir.glob(spectral_type + '*.fits'))
-    return _read_stellar_spectra_path(path)
+        table_copy = lc_table.copy()
+        for band in set(table_copy['band']):
+            band_indices = np.where(table_copy['band'] == band)[0]
+            table_copy['flux'][band_indices] /= self.average_norm_flux(band, pwv[band_indices])
 
-
-@lru_cache()  # Cache I/O
-def get_ref_star_dataframe(spectral_type: str = 'G2') -> pd.DataFrame:
-    """Retrieve PWV values to use as reference values
-
-    Args:
-        spectral_type: Type of reference star (Default 'G2')
-
-    Returns:
-        A DataFrame indexed by PWV with columns for flux
-    """
-
-    rpath = data_paths.stellar_flux_dir / f'{spectral_type}.txt'
-    if not rpath.exists():
-        raise ValueError(
-            f'Data not available for specified star {spectral_type}. '
-            f'Could not find: {rpath}')
-
-    band_names = [f'lsst_hardware_{b}' for b in 'ugrizy']
-    column_names = ['PWV'] + band_names
-    reference_star_flux = pd.read_csv(
-        rpath,
-        sep='\\s',
-        header=None,
-        names=column_names,
-        comment='#',
-        index_col=0,
-        engine='python'
-    )
-
-    for band in band_names:
-        band_flux = reference_star_flux[f'{band}']
-        reference_star_flux[f'{band}_norm'] = band_flux / band_flux.loc[0]
-
-    return reference_star_flux
+        return table_copy
 
 
-def interp_norm_flux(band: str, pwv: Union[Numeric, np.array], spectral_type='G2') -> np.ndarray:
-    """Return normalized reference star flux values
+class VariableCatalog:
+    """A reference star catalog that determines the time dependent PWV concentration from an underlying PWV model"""
 
-    Args:
-        band: Band to get flux for
-        pwv: PWV values to get magnitudes for
-        spectral_type: Type of reference star (Default 'G2')
+    def __init__(self, pwv_model: PWVModel, time_format: str, *spectral_types: str) -> None:
+        """Create a reference star catalog composed of the given spectral types and a PWV model
 
-    Returns:
-        The normalized flux at the given PWV value(s)
-    """
+        Args:
+            *spectral_types: Spectral types for the catalog (e.g., 'G2', 'M5', 'K2')
+        """
 
-    reference_star_flux = get_ref_star_dataframe(spectral_type)
-    if np.any(
-            (pwv < reference_star_flux.index.min()) |
-            (pwv > reference_star_flux.index.max())):
-        raise ValueError('PWV is out of range')
+        self.catalog = ReferenceCatalog(*spectral_types)
+        self.time_format = time_format
+        self.pwv_model = pwv_model
 
-    norm_flux = reference_star_flux[f'{band}_norm']
-    return np.interp(pwv, norm_flux.index, norm_flux)
+    def average_norm_flux(
+            self,
+            band: str,
+            time: Union[float, np.ndarray, Collection],
+            ra: float,
+            dec: float,
+            lat: float = const.vro_latitude,
+            lon: float = const.vro_longitude,
+            alt: float = const.vro_altitude
+    ) -> np.ndarray:
+        """Return the average normalized reference star flux
 
+        Args:
+            band: Band to get flux for
+            time: Time at which the target is observed
+            ra: Right Ascension of the target (Deg)
+            dec: Declination of the target (Deg)
+            lat: Latitude of the observer (Deg)
+            lon: Longitude of the observer (Deg)
+            alt: Altitude of the observer (m)
 
-def average_norm_flux(
-        band: str, pwv: Union[Numeric, np.ndarray], spectral_types: Collection[str] = ('G2', 'M5', 'K2')
-) -> np.ndarray:
-    """Return the average normalized reference star flux
+        Returns:
+            The normalized flux at the given PWV value(s)
+        """
 
-    Args:
-        band: Band to get flux for
-        pwv: PWV values to get magnitudes for
-        spectral_types: Types of reference stars to average over
+        pwv = self.pwv_model.pwv_los(time, ra, dec, lat, lon, alt, time_format=self.time_format)
+        return self.catalog.average_norm_flux(band, pwv)
 
-    Returns:
-        The normalized flux at the given PWV value(s)
-    """
+    def divide_ref_from_lc(
+            self,
+            lc_table: Table,
+            time: Union[float, np.ndarray, Collection],
+            ra: float,
+            dec: float,
+            lat: float = const.vro_latitude,
+            lon: float = const.vro_longitude,
+            alt: float = const.vro_altitude
+    ) -> Table:
+        """Divide reference flux from a light-curve
 
-    return np.average([interp_norm_flux(band, pwv, stype) for stype in spectral_types], axis=0)
+        Recalibrate flux values using the average change in flux of a collection of
+        reference stars.
 
+        Args:
+            lc_table: Astropy table with columns ``flux`` and ``band``
+            time: Time at which the target is observed
+            ra: Right Ascension of the target (Deg)
+            dec: Declination of the target (Deg)
+            lat: Latitude of the observer (Deg)
+            lon: Longitude of the observer (Deg)
+            alt: Altitude of the observer (m)
 
-def divide_ref_from_lc(
-        lc_table: Table, pwv: Union[Numeric, np.ndarray], spectral_types: Collection[str] = ('G2', 'M5', 'K2')
-) -> Table:
-    """Divide reference flux from a light-curve
+        Returns:
+            A modified copy of ``lc_table``
+        """
 
-    Recalibrate flux values using the average change in flux of a collection of
-    reference stars.
-
-    Args:
-        lc_table: Astropy table with columns ``flux`` and ``band``
-        pwv: PWV value to subtract reference star for
-        spectral_types: Type of reference stars to use in calibration
-
-    Returns:
-        A modified copy of ``lc_table``
-    """
-
-    if isinstance(pwv, Collection):
-        if len(pwv) != len(lc_table):
-            raise ValueError('PWV must be a float or have the same length as ``lc_table``')
-
-        pwv = np.array(pwv)
-
-    else:
-        pwv = np.full(len(lc_table), pwv)
-
-    table_copy = lc_table.copy()
-    for band in set(table_copy['band']):
-        band_indices = np.where(table_copy['band'] == band)[0]
-        table_copy['flux'][band_indices] /= average_norm_flux(band, pwv[band_indices], spectral_types)
-
-    return table_copy
+        pwv = self.pwv_model.pwv_los(time, ra, dec, lat, lon, alt, time_format=self.time_format)
+        return self.catalog.divide_ref_from_lc(lc_table, pwv)
